+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.0.40-cloud] - 2024-09-29
 ### Added
 - Cache role membership in edge replication for Atlantis
-- Added scaffolds for add permission event
+- Added scaffolds for add and remove permission event
 
 ### Fixed
 - Fixed assume role when credentials expired